AC-921: Added billing subscription update action
